Check shell status codes
Previously we were executing shell commands and ignored their status codes. With this patch we will check to see if they have a non-success status code and raise an error.

One of the main places this used to show up is if there was a problem downloading something from ruby-lang or if a file was downloaded that didn't exist. When you go to re-run the build script it would see a file existed in the cache and try to use it but fail. Unfortunately since we weren't error checking the failure would come later in code and it was difficult to see where the true failure was coming from.