Add support for Heroku-20
The Dockerfile was based on that for Heroku-18, except with the
`apt-get install` step removed, since:
* `autoconf` and `zip` are already in the base image [1]
* `subversion` is only used when `SVN_URL` is set in the build script,
  which is only required for Ruby < 2.0.

The build scripts were generated using `rake new[<version>,heroku-20]`,
and I added one for each of the latest versions of Ruby 2.4 to 2.7 [2].

I included Ruby 2.4 even though it's no longer maintained [3], since the
EOL date was only 2 weeks ago, and it includes the same security fixes
as all of the other maintained branches (ie: until there is another CVE
announced, it doesn't contain any known vulnerabilities). Including it
will ease the stack upgrade path for customers, given the difficulty
we've seen with the Cedar-14 EOL and customers trying to simultaneously
update stack+Ruby+Rails versions.

It's worth noting that for now, the `rake test` of Heroku-20 will fail,
since the new stack is marked as internal-only, so needs a privileged
API call to create the app with that stack. However I've tested the build
scripts locally.

Refs W-7457565.

[1] https://github.com/heroku/stack-images/blob/master/heroku-20-build/installed-packages.txt
[2] https://www.ruby-lang.org/en/downloads/releases/
[3] https://www.ruby-lang.org/en/downloads/branches/