Switch away from deprecated path-based S3 URLs (#16)
S3 URLs where the bucket name is part of the URL path are deprecated.

Instead, it's recommended to use the virtual-hosted style references,
where the bucket name is part of the domain. 

The latter allows AWS to use DNS to direct requests directly to the
appropriate regions S3 endpoint, rather than having to route via the
global S3 endpoint (which AWS describe as being a single point of
failure/harder to scale etc).

See:
https://github.com/heroku/heroku-buildpack-ruby/pull/1311
https://aws.amazon.com/blogs/aws/amazon-s3-path-deprecation-plan-the-rest-of-the-story/
https://docs.aws.amazon.com/AmazonS3/latest/userguide/VirtualHosting.html
https://docs.aws.amazon.com/general/latest/gr/s3.html#s3_region

GUS-W-11283397.