Refactor + Test + Support preview releases
I've resisted really touching the compilation script because it had no tests. This was tolerable when we ran scripts locally and adjustments could be made on the fly. Now that we're automating them, I wanted to have more confidence in the script. To do that I iteratively extracted, tested, and then refactored code.

In the process I got rid of most branching logic for features we don't support on GitHub actions today. If we need them in the future we have a git history.

The unit tests are fast, but not terribly comprehensive. We could continue down this path of testing and refactoring if we choose, but this was a good enough stoping point for my needs.

The only new behavior added is fixing these bugs that are all centered around preview/rc releases:

- https://github.com/heroku/docker-heroku-ruby-builder/issues/28
- https://github.com/heroku/docker-heroku-ruby-builder/issues/29
- https://github.com/heroku/docker-heroku-ruby-builder/issues/30

While I'm at it I also added a linter (standard).