Add AMD (x86_64) and ARM (aarch64) on Heroku-24 (#38)
* Add AMD (x86_64) and ARM (aarch64) on Heroku-24

AMD/x86_64 is the architecture used for all prior base images. Heroku-24 (Ubuntu 24.04) base image is provided with support for ARM/aarch64 (think m1 Mac or graviton AWS server) support. This PR does several things:

- Introduces support for heroku-24 base image
- Build both arm64 and amd64 architecture binaries

* Use GHA arm runner beta

* Update ruby/setup-ruby to @v1 everywhere

Versions are different across the project, this commit synchronizes them all to v1 which is the latest.

* Move build scripts to bash

We shouldn't need Ruby installed on the runner, and again the container just to be able to build Ruby versions. This commit moves functionality that doesn't **need** to be in Ruby to bash. This is needed as ruby/setup-ruby does not currently work with ARM https://github.com/ruby/setup-ruby/issues/577.

* Manually install docker

* Remove Ruby requirement on GHA for building binaries

Remove the Ruby requirement in the local runtime to build rubies. It is still needed on the stack image, but this allows us to decouple from ruby/setup-ruby which is itself blocked on building binaries. We still need it for running unit tests on CI.

- Remove Rake and Rakefile
- Update README to use bash commands
- Update all ci.yml and build_ruby.yml to use bash instead of Ruby

* Apply suggestions from code review

Co-authored-by: Colin Casey <casey.colin@gmail.com>

* Fix linting check in CI

* Remove docker file from root and ignore

It is copied in to root as needed and shouldn't be checked in.

* Apply suggestions from code review

Co-authored-by: Colin Casey <casey.colin@gmail.com>

---------

Co-authored-by: Colin Casey <casey.colin@gmail.com>