Remove map_error from code
From https://www.reddit.com/r/rust/comments/1dkhl0n/my_rust_error_handling_best_practices/ (full text below). Which links to this video https://www.youtube.com/watch?v=j-VQCYP7wyw&t=1s. The overall goal of this commit is to reduce complexity and code.

We can remove the "full" error handling i.e. wrapping every error in a custom enum because this is a "script" where "ease of editing" and "speed to read" are more important than long term extensibility and error-correctness. Instead we can use `Box<dyn std::error::Error>`. 

We can choose to "upgrade" error handling later using the progressive steps listed below:

## Text from linked url

```
Here are my error best practices I have matured to. This is working pretty well for me.

5 points summary: (but watch the video for the full context and examples)

For tests & examples, Box<dyn Error>. Zero dependency, simple, almost as flexible as Anyhow, but best of all "progressive" with the production way to do the ? and error handling.

Next for production code, I have Error as enum, with a Custom(String) variant, with the from &str and from String, and I start to build other variants.

As the code matures, I remove the Error::Custom(String) and make sure all is fully typed.

Also, very important, I make sure the WHOLE meaning is captured in the variant name and properties name. (The display is mostly for command line utilities.)

I like to use derive_more rather than thiserror to remove the From/Display boilerplate. It allows me to use the From, and wire the debug as the display for non-command line errors (as they will be captured in a structured way for further rendering anyway).

The best of this approach is that it is very progressive from 1...5, and I do not expose any "lib" way to handle Error, as all of the Errors are standard Rust Errors.

There is no absolute truth, but just wanted to share what is working well for me.
```