Hybrid Rust (outside) + Bash (inside) (#47)
* Rewrite it in Rust (RiiR)

Rewrites the project from Bash to Rust (previously was rewritten from Ruby to Bash).

## Why Rust?

I bet you're thinking "Why not something simpler like bash or Ruby?" This library was originally written in Ruby and shelled out. That caused bootstrapping problems, for example when rolling out ARM support, the github action for installing Ruby did not yet support ARM so we had to re-write the logic in Bash (or bootstrap our own version of Ruby with bash). We chose to re-write the library in bash. So why not keep it in bash? Even though bash doesn't need "bootstrapping" authors rely on system tools and packages which may or may not already be installed, and may or may not vary between operating systems. For example GNU grep uses different arguments than BSD (mac) grep. So while there's not a "bash bootstrapping problem" installing dependencies and ensuring scripts work across multiple platforms is tricky. It's easy to write quick scripts, but hard to maintain and do well.

Don't you have a Rust bootstrapping problem now? As of Ruby 3.2 YJIT requires a Rust toolchain to support the `--enable-yjit` config flag, so Rust is already a requirement of a full Ruby install. That means that even if we didn't write our scripts in Rust, we still need to have it available on the system when we build Ruby anyway. It's also historically been an easy to install language.

* Use proper java properties parsing

Turns out, we don't need to re-write java properties parsing logic, an existing crate exists.

* Prefer `expect` over `unwrap`

* Ensure directory exists before writing tar dir

* Fix CI

* Split long command to multiple lines

* Update build action, introduce jruby_build action

* Standardize error reporting

* Clean up env var representation

* GHA commands on one line

* Hybrid workflow

Outer scripts are still Rust, inner script is now bash

* Clean up bash script

- Reduce logic duplication by moving the `./configure` options to an array
- Add some minimal comments
- Prettier command logging via trap instead of `set -o trace`

* Rename `inside_docker` to `shared`

Previously this held the `make_ruby.rs` script and a shared set of tools. Now it only holds those tools.

* Reduce/remove expect() calls

* Workflow commands on one line

* Update docs