Remove version check (#90)
Ruby 4.0.0-preview2 cannot compile on Heroku-22 due to this incorrect version check. It is meant to guard against trying to compile a version of Ruby that uses an old version of Open SSL due to https://bugs.ruby-lang.org/issues/18658.

All versions of Ruby that are supported by Ruby core work with heroku-22 and heroku-24. This check is no longer needed.